leaudio: Improve handling Autonomous Release
With this patch, when Autonomous release is detected, state machine
notifies upper layer that group is Releasig and also sets group state
and a Target state.

Also with this patch, code will make sure to not confuse device which is
attaching to the stream with the one which is just in the Releasing
state - this could lead to CIG being to removed when group gets IDLE
state.

Bug: 348654441
Bug: 347161829
Flag: Exempt, regression tested with unit test, new test added
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
(Cherry-picked Change-Id: Iea84f6974fcdc9f58f5346269c758dade238f112)

CRs-Fixed: 4103202
Change-Id: Ieec2009b7476ed15d6beecd382ef390ff4d481f9