Enhancements for multi LE codecs switching supported[3/3]
Use-case:
1. Connected headset, selected codec as Aptx_LE as default.
2. Switched codec to LC3.
3. Rebooted BT
4. Reconnected headset, codec backed to Aptx_LE.

Root cause: The preferred config info was reset when BT turn off.
So it set Apxt_LE back when reconnection as default.

Fixed: If the remote supports multi LE codecs, store it
in bluetooth_leaudio_codec_map. Mark its value as true when
codec is not LC3. It sets LC3 when it detects the value is
true after BT reboot.

Change-Id: Ie762ade9077fb9aff3163b6f1b5d3b41778ef68a
CRs-Fixed: 4086166