xpan: remove the Lex device after ACL disconnects instead of GATT disconnects
Currently, the Lex device is added to lexAvailableTransportDevices_ in
OnVscEvent, which is sent by QHCI after the ACL connection is
established and the QLL feature is read. However, the Lex device is
removed from lexAvailableTransportDevices_
when the GATT channel of the LEA client is disconnected.
This means that adding and removing the Lex device is not symmetric.
If the GATT channel is disconnected and then reconnected while the ACL
remains connected, the Lex device will not be re-added to
lexAvailableTransportDevices_, leading to a streaming_start failure.

Change-Id: Ic9c39cb487f56263f53e67a9bf0e57c8fd79ad01
CRs-Fixed: 3987360