ADM: SHO between 2 devices is failing.
When SHO between 2 devices is performed, the active device remain
unchanged because of the fallback mechanism which was recently
added during the active device change.
This is causing the already existing BREDR device as always active.

Removed the fallback device function calling, as it can be triggered
during device disconnection and not during the active device change.

CRs-Fixed: 4030113
Change-Id: Ie564b285ef3684ed02bae205373946a28e1774bb