SHO: SHO failing issue with LE devices is fixed.
When set active null is being called to the LE stack,
active device change call back is not updated to app,
causing the UI to show 2 active devices.

The inactive status update to app is deffered until the state
is transitioned to IDLE. But in recent changes, due to
caching mechanism, codec configured state is also treated
as idle state and thus causing the group stream status to not
move to IDLE state, leading to not sending inactive update to app.

CRs-Fixed: 4060809
Change-Id: I1b1cd774399a7eab524b3ef59f2f1c901507aa70