Remove MessageQueue reflection from Bluetooth's copy of TestLooper
In ag/31652888 I removed the use of MessageQueue reflection
from TestLooper. This is to unblock the new MessageQueue
implementation.
In aosp/3473844, Bluetooth made a private copy of the old code.
This change is to Bluetooth's copy up to speed.

Ignore-AOSP-First: depends on TestLooperManager changes in internal that haven't merged to AOSP yet
Flag: android.os.message_queue_testability
Bug: 379472827 
Test: atest GoogleBluetoothInstrumentationTests
Change-Id: I2c786badc34c4f7dd46d1ae0f5000fe5702de10a