SystemServer: race between enableBle and apm
If apm spawn two event back to back, the second will be queued up to
when Bluetooth reach a stable state (BLE_ON or ON or OFF). But as soon
as Bluetooth exit BLE_ON, some apps may ask to restart the BLE part.
This event is handled and because the apm event is delayed on a timer,
it will be treated at a later date.

To prevent this, we are cancelling the queued apm event and executing it
directly once reaching OFF, before to generate a "restart" event

Bug: 402563502
Bug: 402332571
Test: atest ServiceBluetoothTests
Flag: com.android.bluetooth.flags.enable_ble_while_disabling_airplane
Change-Id: I2c22468e82404618f75eb2576c83c5231b004278