Don't remove bond on encryption failure in peripheral role
Encryption failure in LE peripheral role generates a SMP_COMPLT_EVT event with an error. This may result in the bond getting removed.
If the link could not be encrypted in peripheral role with a bonded device, we must just disconnect the link and not remove the bond.

Test: mmm packages/modules/Bluetooth
Test: Manual | Keep reconnecting Android Wear devices, there must not be any bond loss
Flag: com.android.bluetooth.flags.le_peripheral_enc_failure
Bug: 402207015
Bug: 403313352
Change-Id: I98e86140d3cc5b62b027c03d8fb8135401da93c0