leaudio: Clear data path state when it is failed to be created
Make sure that stack has valid data path state when it is failed to be
set (usually due to controller issue).
If the state is incorrect, stack will try to remove data path when CIS
is disconnected which will lead to assert in btm_iso layer.

Bug: 404423428
Flag: Exempt, day to day fix
Test: atest bluetooth_le_audio_test
Change-Id: I311a86cfe54ccba39ae697bdc85a53ff254e44af