SHO: SHO betweeen LE devices during streaming.
When Streaming is already in progress, the LE device switch
is failing because of refactoring changes brought in as a part
of caching the ASE state to codec_configured instead of idle.
As a part of the refactoring, device disconnection happening during
the configure streaming request due to non cancellation of
timer set for acheving the codec configured state. And active status
is also not updated to application layer due to non clearance of deferred varaibles.

CRs-fixed: 4119778
Change-Id: I1b1cd774399a7eab524b3ef59f1f2c901507a7aa
(cherry picked from commit 05db04e43249884a86d5984e5adeb7ab8c801cd7)