Disalow setActive when SCO is in opening state
When more than one remote device is connected to DUT,
and setActive requests are received while SCO is in setup,
the SCO state does not align with the event received, resulting in
the events getting ignored.
SCO should either be connected or disconnected for an
another setActive to process.
Hence, setActive should fail if SCO is in opening state.

CRs-Fixed: 4134932
Change-Id: I8e5c957210f847d4b5f7662c3309046273dac2a7