BTIF: Change to ensure l2cap data write call from GATT tool takes place on main thread.
When L2cap data write is called from Bluetooth certification tools, data write is performed on the thread ran by the tool. However, NOCP packets are received on Bluetooth main thread, which causes the next set of packets to be sent on a different thread (main thread). This causes race condition, leading to buffer corruption and hence stack crash.

To avoid this, change is made to send data on main thread, to ensure data is sent on a single thread.

CRs-Fixed: 4376412

Change-Id: I6ddd27bcd968a394a03f5b1c2dde3196ca9dc47c