HFP: Fixed the ANR issue caused by deadlock of mutexes
When the subscription changes, Thread#1 calls stopListenForPhoneState,
acquires the mPhoneStateListenerLock and then waits for the
mDeviceEventMap lock. At the same time, when Bluetooth turn off, the
HeadsetPhoneState.cleanup in thread#2 context first acquires the
mDeviceEventMap lock and then attempts to acquire the
mPhoneStateListenerLock, resulting in a deadlock.

Fix: Remove the lock dependency.
stopListenForPhoneState does not need to be invoked while holding the
mDeviceEventMap lock.

CRs-Fixed: 4371605
Change-Id: If0ced852a25edfa90a42d30184b309197be608e9